backend/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
werkzeug
joblib
numpy
scikit-learn
PyJWT
reportlab
psycopg2-binary
flask-sqlalchemy
